+++ cdk/src/open-data-platform/data-plane/data-import/write-water-systems-data-handler.handler.ts
@@ -1,6 +1,7 @@
 import { RDSDataService } from 'aws-sdk';
 import { BatchExecuteStatementRequest, SqlParametersList } from 'aws-sdk/clients/rdsdataservice';
 import { geoJsonHandlerFactory } from './handler-factory';
+import { WaterSystemsTableRowBuilder } from '../model/water-systems';
 
 // As of 2022-06-27, this should have 26010 rows.
 const s3Params = {
@@ -9,112 +10,6 @@ const s3Params = {
 };
 
 const SCHEMA = 'public';
-
-/**
- * Single row for water systems table.
- */
-class WaterSystemsTableRow {
-  // Field formatting conforms to rows in the db. Requires less transformations.
-
-  // Water system identifier.
-  pws_id: string;
-  // Water system name.
-  pws_name: string;
-  // Reported or estimated number of lead pipes in the boundary.
-  lead_connections_count: number;
-  // Reported or estimated number of connections in the boundary.
-  service_connections_count: number;
-  // Number of people served by the water system.
-  population_served: number;
-  // GeoJSON representation of the boundaries.
-  geom: string;
-
-  constructor(
-    pws_id: string,
-    pws_name: string,
-    lead_connections_count: number,
-    service_connections_count: number,
-    population_served: number,
-    geom: string,
-  ) {
-    this.pws_id = pws_id;
-    this.pws_name = pws_name;
-    this.lead_connections_count = lead_connections_count;
-    this.service_connections_count = service_connections_count;
-    this.population_served = population_served;
-    this.geom = geom;
-  }
-}
-
-/**
- * Builder utility for rows of the water systems table.
- */
-class WaterSystemsTableRowBuilder {
-  private readonly _row: WaterSystemsTableRow;
-
-  constructor() {
-    this._row = new WaterSystemsTableRow('', '', 0, 0, 0, '');
-  }
-
-  pwsId(pwsId: string): WaterSystemsTableRowBuilder {
-    this._row.pws_id = pwsId;
-    return this;
-  }
-
-  pwsName(pwsName: string): WaterSystemsTableRowBuilder {
-    this._row.pws_name = pwsName;
-    return this;
-  }
-
-  geom(geom: string): WaterSystemsTableRowBuilder {
-    this._row.geom = geom;
-    return this;
-  }
-
-  leadConnectionsCount(leadConnectionsCount: number): WaterSystemsTableRowBuilder {
-    this._row.lead_connections_count = leadConnectionsCount;
-    return this;
-  }
-
-  serviceConnectionsCount(serviceConnectionsCount: number): WaterSystemsTableRowBuilder {
-    this._row.service_connections_count = serviceConnectionsCount;
-    return this;
-  }
-
-  populationServed(populationServed: number): WaterSystemsTableRowBuilder {
-    this._row.population_served = populationServed;
-    return this;
-  }
-
-  build(): SqlParametersList {
-    return [
-      {
-        name: 'pws_id',
-        value: { stringValue: this._row.pws_id },
-      },
-      {
-        name: 'pws_name',
-        value: { stringValue: this._row.pws_name },
-      },
-      {
-        name: 'lead_connections_count',
-        value: { doubleValue: this._row.lead_connections_count },
-      },
-      {
-        name: 'service_connections_count',
-        value: { doubleValue: this._row.service_connections_count },
-      },
-      {
-        name: 'population_served',
-        value: { doubleValue: this._row.population_served },
-      },
-      {
-        name: 'geom',
-        value: { stringValue: this._row.geom.toString() },
-      },
-    ];
-  }
-}
 
 /**
  *  Writes rows into the water systems table.
